远程数据库操作测试成功

+++ src/main/java/xmu/oomall/mapper/DefaultFreightMapper.java
@@ -8,5 +8,5 @@ import xmu.oomall.domain.DefaultFreightPo;
 @Component
 public interface DefaultFreightMapper {
 
-    DefaultFreightPo selectByPrimaryKey(Long id);
+    DefaultFreightPo selectByPrimaryKey(Integer id);
 }
